+++ SBBMobileDesignSystemSwiftUIDemo/Views/DetailViews/BannerDemo.swift
@@ -8,19 +8,20 @@ import SBBMobileDesignSystemSwiftUI
 struct BannerDemo: View {
     
     @Binding var colorScheme: ColorScheme
-    @EnvironmentObject var model: BannerViewModel
+    @Binding var selectedBanner: SBBEnvironmentBanner
     
     var body: some View {
         VStack(alignment: .center, spacing: 16) {
-            SBBFormGroup(title: "Banner type") {
-                SBBRadioButton(isOn: $model.none, label: "None")
-                SBBRadioButton(isOn: $model.dev, label: "dev")
-                    .sbbEnvironmentBanner(.dev)
-                SBBRadioButton(isOn: $model.test, label: "test")
+            SBBRadioButtonGroup(title: "Banner type", selection: $selectedBanner, tags: SBBEnvironmentBanner.allCases) {
+                SBBRadioButton(label: "None")
+                    .sbbEnvironmentBanner(.none)
+                SBBRadioButton(label: "test")
                     .sbbEnvironmentBanner(.test)
-                SBBRadioButton(isOn: $model.int, label: "int")
+                SBBRadioButton(label: "dev")
+                    .sbbEnvironmentBanner(.dev)
+                SBBRadioButton(label: "int")
                     .sbbEnvironmentBanner(.int)
-                SBBRadioButton(isOn: $model.poc, label: "poc", showBottomLine: false)
+                SBBRadioButton(label: "poc", showBottomLine: false)
                     .sbbEnvironmentBanner(.poc)
             }
             Spacer()
@@ -35,9 +36,8 @@ struct BannerDemo: View {
 struct BannerDemo_Previews: PreviewProvider {
     static var previews: some View {
         Group {
-            BannerDemo(colorScheme: .constant(.light))
-            BannerDemo(colorScheme: .constant(.dark))
+            BannerDemo(colorScheme: .constant(.light), selectedBanner: .constant(.none))
+            BannerDemo(colorScheme: .constant(.dark), selectedBanner: .constant(.none))
         }
-            .environmentObject(BannerViewModel())
     }
 }